% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeCoxSurv.R
\name{imputeCoxSurv}
\alias{imputeCoxSurv}
\title{Fit cox survival to all variants in a SummarizedExperiment object}
\usage{
imputeCoxSurv(se, time, event, covariates)
}
\arguments{
\item{se}{SummarizedExperiment object as generated with `readImputeGds` and `addCov` functions 
or includes 1 assay containing allele dosages; rowData with SNP information;
colData including time, event and covariates.}

\item{time}{character(1) string that matches time column name in pheno.file}

\item{event}{character(1) string that matches event column name in pheno.file}

\item{covariates}{character vector with matching column names in pheno.file of covariates of interest}
}
\value{
Generates se where survival results are kept in rowData.
}
\description{
Performs survival analysis using Cox proportional hazard models on imputed genetic data stored in SummarizedExperiment object
}
