% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/michiganCoxSurv.R
\name{michiganCoxSurv}
\alias{michiganCoxSurv}
\title{Fit cox survival to all variants in a .vcf.gz file from Michigan imputation server}
\usage{
michiganCoxSurv(vcf.file, pheno.file, time.to.event, event, covariates,
  sample.ids, output.name, chunk.size, info.filter, maf.filter,
  verbose = TRUE)
}
\arguments{
\item{vcf.file}{character(1) path to corresponding info file.}

\item{pheno.file}{matrix(1) comprising phenotype data.}

\item{time.to.event}{character(1) string that matches time column name in pheno.file}

\item{event}{character(1) string that matches event column name in pheno.file}

\item{covariates}{character vector with matching column names in pheno.file of covariates of interest}

\item{sample.ids}{character vector with sample ids to include in analysis}

\item{output.name}{character(1) string with output name}

\item{chunk.size}{integer(1) number of variants to process per thread}

\item{info.filter}{integer(1) of imputation quality score filter (i.e. 0.7 will filter info > 0.7)}

\item{maf.filter}{integer(1) filter out minor allele frequency below threshold (i.e. 0.005 will filter MAF > 0.005)}

\item{verbose}{logical(1) for messages that describe which part of the analysis is currently being run}
}
\value{
Saves text file directly to disk that contains survival analysis results
}
\description{
Performs survival analysis using Cox proportional hazard models on imputed genetic data stored in compressed VCF files
}
\examples{
vcf.file <- system.file(package="gwasurvivr","extdata", "sanger.pbwt_reference_impute.vcf.gz")
pheno.fl <- system.file(package="gwasurvivr", "extdata", "simulated_pheno.txt")
pheno.file <- read.table(pheno.fl, sep=" ", header=TRUE, stringsAsFactors = FALSE)
library(tidyverse)
library(magrittr)
pheno.file <- pheno.file \%>\% 
                   mutate(SexFemale=case_when(sex=="female"~1L,
                          sex=="male"~0L)) \%>\% 
                   select(-ID_1)
sample.ids <- pheno.file \%>\%
                   filter(group=="experimental") \%$\%
                   ID_2 

}
