% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfCoxSurv.R
\name{vcfCoxSurv}
\alias{vcfCoxSurv}
\title{Fit cox survival to all variants in a VCF file}
\usage{
vcfCoxSurv(vcf.file, info.file, MAF, Rsq, chunk.size, pheno.file, time, event,
  covariates, sample.ids, output.name, verbose = TRUE)
}
\arguments{
\item{vcf.file}{character(1) path to VCF file.}

\item{chunk.size}{integer(1) number of variants to process per thread}

\item{pheno.file}{numeric matrix(1) comprising phenotype data.}

\item{time}{character(1) string that matches time column name in pheno.file}

\item{event}{character(1) string that matches event column name in pheno.file}

\item{covariates}{character vector with matching column names in pheno.file of covariates of interest}

\item{sample.ids}{character vector with sample ids to include in analysis}

\item{output.name}{character(1) string with output name}

\item{vcf.file}{character(1) path to corresponding info file.}
}
\value{
Saves text file directly to disk that contains survival analysis results
}
\description{
Performs survival analysis using Cox proportional hazard models on imputed genetic data stored in compressed VCF files
}
\examples{
set.seed(222)
vcf.file <- system.file(package="SurvivR", "extdata", "chr21.25000005-25500000.vcf.gz")
fl <- system.file(package="SurvivR", "extdata", "pheno_file.txt") 
pheno.file <- read.table(fl, sep="\\t", header = T)
time <- "intxsurv_1Y"
event <- "dead_1Y" 
covariates <- c("distatD", "age") 
sample.ids = sample(rownames(pheno.file), size=190)
pheno.file$distatD <- as.integer(pheno.file$distatD) - 1L
pheno.file <- as.matrix(pheno.file)
output.name <- tempfile()
vcfCoxSurv(vcf.file, chunk.size, pheno.file, time, event, covariates, sample.ids, output.name)
 
}
