% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute2CoxSurv.R
\name{impute2CoxSurv}
\alias{impute2CoxSurv}
\title{Fit cox survival to all variants from a standard IMPUTE2 output after genotype imputation}
\usage{
impute2CoxSurv(impute.file, sample.file, chr, covariate.file,
  id.column = NULL, sample.ids = NULL, time.to.event, event, covariates,
  interaction = NULL, print.covs = "only", out.file, maf.filter = NULL,
  info.filter = NULL, flip.dosage = TRUE, verbose = TRUE,
  clusterObj = NULL)
}
\arguments{
\item{impute.file}{character(1) of IMPUTE2 file}

\item{sample.file}{character(1) of sample file affiliated with IMPUTE2 file}

\item{chr}{numeric(1) denoting chromosome number}

\item{covariate.file}{data.frame(1) or matrix(1) comprising phenotype information; first column is required to be sample IDs}

\item{sample.ids}{character(1) vector of sample IDs to keep in survival analysis}

\item{time.to.event}{character(1) of column name in covariate.file that represents the time interval of interest in the analysis}

\item{event}{character(1) of column name in covariate.file that represents the event of interest to be included in the analysis}

\item{covariates}{character(1) vector with exact names of columns in covariate.file to include in analysis}

\item{out.file}{character(1) of output file name (do not include extension)}

\item{maf.filter}{numeric(1) to filter minor allele frequency (MAF)}

\item{info.filter}{numeric(1) to filter imputation INFO score (i.e. choosing 0.7 means filtering info>0.7)}

\item{flip.dosage}{logical(1) to flip which allele the dosage was calculated on, default=TRUE}

\item{verbose}{logical(1) for messages that describe which part of the analysis is currently being run}
}
\value{
Saves text file directly to disk that contains survival analysis results
}
\description{
Performs survival analysis using Cox proportional hazard models on imputed genetic data from IMPUTE2 output
}
\examples{
library(tidyverse)
library(magrittr)
impute.file <- system.file(package="gwasurvivr","extdata","impute_example.impute2")
sample.file <- system.file(package="gwasurvivr",
                           "extdata", 
                          "impute_example.impute2_sample")
chr <- 14
covariate.file <- system.file(package="gwasurvivr", 
                              "extdata",
                             "simulated_pheno.txt")
covariate.file <- read_delim(covariate.file, delim=" ")
covariate.file <- covariate.file \%>\% 
   mutate(SexFemale=case_when(sex=="female"~1L,
                              sex=="male"~0L)) \%>\%
   select(-ID_1)
covariate.file \%>\% head
sample.ids <- covariate.file \%>\%
   filter(group=="experimental") \%$\%
   ID_2 
impute2CoxSurv(impute.file=impute.file,
              sample.file=sample.file,
              chr=14,
              covariate.file=covariate.file,
              sample.ids=sample.ids,
              time.to.event="time",
              event="event",
              covariates=c("age", "SexFemale", "bmiOVWT"),
              out.file="impute_example",
              maf.filter=0.01,
              info.filter=0.7,
              flip.dosage=TRUE,
              verbose=TRUE)  
 
}
