% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute2CoxSurv.R
\name{impute2CoxSurv}
\alias{impute2CoxSurv}
\title{Fit cox survival to all variants from a standard IMPUTE2 output after genotype imputation}
\usage{
impute2CoxSurv(impute.file, sample.file, chr, covariate.file, id.column,
  sample.ids = NULL, time.to.event, event, covariates, inter.term = NULL,
  print.covs = "only", out.file, chunk.size = 10000, info.filter = 0.3,
  maf.filter = 0.05, flip.dosage = TRUE, verbose = TRUE,
  clusterObj = NULL)
}
\arguments{
\item{impute.file}{character(1) of IMPUTE2 file}

\item{sample.file}{character(1) of sample file affiliated with IMPUTE2 file}

\item{chr}{numeric(1) denoting chromosome number}

\item{covariate.file}{data.frame(1) comprising phenotype information, all covariates to be added in the model must be numeric.}

\item{id.column}{character(1) giving the name of the ID column in covariate.file.}

\item{sample.ids}{character(1) vector of sample IDs to keep in survival analysis}

\item{time.to.event}{character(1) of column name in covariate.file that represents the time interval of interest in the analysis}

\item{event}{character(1) of column name in covariate.file that represents the event of interest to be included in the analysis}

\item{covariates}{character(1) vector with exact names of columns in covariate.file to include in analysis}

\item{inter.term}{character(1) string giving the column name of the covariate that will be added to the interaction term with SNP (e.g. \code{term*SNP}). See details.}

\item{print.covs}{character(1) string of either \code{"only"}, \code{"all"} or \code{"some"}, defining which covariate statistics should be printed to the output. See details.}

\item{out.file}{character(1) of output file name (do not include extension)}

\item{chunk.size}{integer(1) number of variants to process per thread}

\item{info.filter}{numeric(1) to filter imputation INFO score (i.e. choosing 0.7 means filtering info>0.7). Default is `NULL`, no filtering is set.}

\item{maf.filter}{numeric(1) to filter minor allele frequency (i.e. choosing 0.05 means filtering MAF>0.05). User can set this to `NULL`` if no filtering is preffered. Default is 0.05.}

\item{flip.dosage}{logical(1) to flip which allele the dosage was calculated on, default=TRUE}

\item{verbose}{logical(1) for messages that describe which part of the analysis is currently being run}

\item{clusterObj}{A cluster object that can be used with the \code{parApply} function. See details.}
}
\value{
Saves text file directly to disk that contains survival analysis results.
}
\description{
Performs survival analysis using Cox proportional hazard models on imputed genetic data from IMPUTE2 output
}
\details{


Testing for SNP-covariate interactions:          
User can define the column name of the covariate that will be included in the interaction term. 
For example, for given covariates \code{a} and \code{b}, where \code{c} is defined as the \code{inter.term} the model will be:
\code{~ a + b + c + SNP + c*SNP}.

Printing results of other covariates:       
\code{print.covs} argument controls the number of covariates will be printed as output. The function is set to \code{only}
by default and will only print the SNP or if an interaction term is given, the results of the interaction 
term (e.g. \code{SNP*covariate}). Whereas,  \code{all} will print results (coef, se.coef, p.value etc) of all covariates 
included in the model. \code{some} is only applicable if an interaction term is given and will print the results for SNP, 
covarite tested for interaction and the interaction term. User should be mindful about using the \code{all} option, as
it will likely slow down the analysis and will increase the output file size. 

User defined parallelization:  
This function uses \code{parApply} from \code{parallel} package to fit models to SNPs in parallel. 
User is not required to set any options for the parallelization. 
However, advanced users who wish to optimize it, can provide a cluster object generated by \code{makeCluster} 
family of functions that suits their need and platform.
}
\examples{
library(gwasurvivr)
library(tidyverse)
library(magrittr)
impute.file <- system.file(package="gwasurvivr","extdata","impute_example.impute2")
sample.file <- system.file(package="gwasurvivr",
                           "extdata", 
                          "impute_example.impute2_sample")
chr <- 14
covariate.file <- system.file(package="gwasurvivr", 
                              "extdata",
                             "simulated_pheno.txt")
covariate.file <- read_delim(covariate.file, delim=" ")
covariate.file <- covariate.file \%>\% 
   mutate(SexFemale=case_when(sex=="female"~1L,
                              sex=="male"~0L)) \%>\%
   select(-ID_1)
covariate.file \%>\% head
sample.ids <- covariate.file \%>\%
   filter(group=="experimental") \%$\%
   ID_2 
impute2CoxSurv(impute.file=impute.file,
              sample.file=sample.file,
              chr=14,
              covariate.file=covariate.file,
              id.column="ID_2",
              sample.ids=sample.ids,
              time.to.event="time",
              event="event",
              covariates=c("age", "SexFemale", "DrugTxYes"),
              inter.term=NULL,
              print.covs="only",
              out.file="impute_example",
              chunk.size=10000,
              maf.filter=0.01,
              info.filter=0.7,
              flip.dosage=TRUE,
              verbose=TRUE,
              clusterObj=NULL)  
 
}
